fix a lot of types

+++ firmware/u4a2/user/usr_grey.c
@@ -79,7 +79,7 @@ void UserGreyInit(byte usrGreyHandler) {
  * Note:            None
  *****************************************************************************/
 void UserGreyProcessIO(void) {
-    if ((usb_device_state < CONFIGURED_STATE) || (UCONbits.SUSPND == 1)) return;
+    if ((usb_device_state < CONFIGURED_STATE) || (UCONbits.SUSPND == (unsigned) 1)) return;
     /* here enter the code that want to be called periodically,
      * per example interaction with buttons and leds */
 }/*end UserGreyProcessIO*/
@@ -149,9 +149,9 @@ void UserGreyReceived(byte* recBuffPtr, byte len, byte handler) {
             break;
     }/*end switch(s)*/
 
-    if (userGreyCounter != 0) {
+    if (userGreyCounter != (byte) 0) {
         j = 255;
-        while (mUSBGenTxIsBusy() && j-- > 0); /* pruebo un maximo de 255 veces */
+        while (mUSBGenTxIsBusy() && j-- > (byte) 0); /* pruebo un maximo de 255 veces */
             if (!mUSBGenTxIsBusy())
                 USBGenWrite2(handler, userGreyCounter);
     }/*end if*/
