+++ firmware/u4a2/user/usr_motors.h
@@ -25,8 +25,8 @@
 #define TIME_UNIT        2000
 #define LONG_TIME_UNIT   5000
 
-#define C_ID_MOTORS 253
-#define C_TRIES 2
+#define C_ID_MOTORS   (byte) 253
+#define C_TRIES       (byte) 2
 
 /** S T R U C T U R E S ******************************************************/
 typedef union MOTORS_DATA_PACKET {
